rag-api/requirements.txt
fastapi>=0.115
uvicorn>=0.32
google-genai>=1.0
qdrant-client>=1.12
pandas>=2.2
openpyxl>=3.1
pydantic>=2.0
httpx>=0.28
